Use decimal instead of double

+++ src/HackathonProblem/Service/Hr/Director/Calculator/HarmonicMeanCalculator.cs
@@ -2,10 +2,10 @@ namespace HackathonProblem.Service.Hr.Director.Calculator;
 
 public static class HarmonicMeanCalculator
 {
-    public static double Calculate(IEnumerable<int> satisfactions)
+    public static decimal Calculate(IEnumerable<int> satisfactions)
     {
         var readOnlySatisfactions = satisfactions.ToList().AsReadOnly();
-        
-        return readOnlySatisfactions.Count / readOnlySatisfactions.Sum(satisfaction => 1.0 / satisfaction);
+
+        return readOnlySatisfactions.Count / readOnlySatisfactions.Sum(satisfaction => 1.0m / satisfaction);
     }
 }
